added menu entity

+++ src/main/java/com/company/food/dto/FoodCr.java
@@ -3,12 +3,13 @@ package com.company.food.dto;
 import lombok.Builder;
 
 import javax.validation.constraints.NotBlank;
+import java.math.BigDecimal;
 
 public record FoodCr(
         @NotBlank(message = "{valid.food.name}")
         String name,
 
         @NotBlank(message = "{valid.food.price}")
-        Double price
+        BigDecimal price
 ) {
 }
