+++ src/main/java/com/company/food/dto/FoodResp.java
@@ -2,11 +2,12 @@ package com.company.food.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 @Builder
 public record FoodResp(
         UUID id,
         String name,
-        Double price
+        BigDecimal price
 ) {
 }
